Added full windows support
